Changed filename + created base class for slideshow ws test classes
